Added a sub module called generic_parser
This is actually a clone of the Wordpress parser with one important difference. It uses cURL to get the file instead of simplexml_load_file. This is so that I could add the curl_setopt to use the proxy server. So this parser can get feeds from off campus while the others cannot.