Update dependencies and improve logging in the live_feeds module
This commit makes several changes to prepare the Live Feeds module for a new release. It expands the supported versions of the drupal/date_ap_style dependency to include the 2.0 release. It also enhances the GetFeed service to use a logger service, so any failed requests to retrieve feeds can be more effectively logged. Other code readability and organization improvements included as well.